requirements.txt
flask==3.0.3
python-dotenv==1.0.1
mysql-connector-python==9.0.0
flask-cors==4.0.1
flask-limiter==3.8.0
flask-talisman==1.1.0
PyJWT==2.9.0
